requirements.txt
torch>=1.12.0
torch-geometric
transformers>=4.30.0
rdkit-pypi
pandas
numpy
scikit-learn
scipy
xgboost
tqdm